test2

+++ app/Http/Controllers/PostController.php
@@ -27,7 +27,7 @@ class PostController extends Controller
     }
 
 
-
+ 
 
    public function show($id){
     $user= User::findOrFail($id);
@@ -62,17 +62,18 @@ class PostController extends Controller
           
            $addpost->save();
            
-          Session::flash('success','Your post has now been published');
+          // Session::flash('success','Your post has now been published');
 
-           return redirect('/posts'); //here supposed to be a specific user posts "user profile"
+           return redirect('/posts')->with('success','Your post has now been published'); //here supposed to be a specific user posts "user profile"
        
 
     }
 
 
     public function edit($id){
+
         $post = Post::findOrFail($id);
-        //dd($id);
+         $this->authorize('edit',$post);
         return view("editpost",['post'=>$post]); 
         
     }
@@ -91,8 +92,8 @@ class PostController extends Controller
       $updatepost->content = request("content");
       $updatepost->image = $request->file('image')->store('/images','public');
       $updatepost->save();
-       Session::flash('message','Your post has now been updated');
-      return redirect('/posts');
+       // Session::flash('message','Your post has now been updated');
+      return redirect('/posts')->with('success','Your post has now been updated');
 
     }
 
@@ -101,9 +102,10 @@ class PostController extends Controller
     public function destroy($id){
         
         $post = Post::findOrFail($id);
+        $this->authorize('destroy',$post);
         $post->delete();
-         Session::flash('remove','Your post has now been removed');
-        return redirect('/posts');
+         // Session::flash('remove','Your post has now been removed');
+        return redirect('/posts')->with('success','Your post has now been removed');
     }
 
 }
